update: separate generic labels mixin
