+++ src/mixins/measurement.js
@@ -517,13 +517,13 @@ export const MeasurementMixin = (superclass) =>
             const label = this.createLabelSprite(
                 `${distance.toFixed(3)}Å`,
                 `label-for-${distance}`,
+                { scale: 0.75 },
             );
             const atomConnections = this.atomConnections.children;
             const line = atomConnections[atomConnections.length - 1];
             line.userData.label = label;
             label.position.set(...line.geometry.boundingSphere.center);
             label.visible = true;
-            label.scale.set(0.75, 0.75, 0.75);
             this.measurementLabels.add(label);
             this.scene.add(this.measurementLabels);
             this.render();
