+++ src/wave.js
@@ -5,18 +5,16 @@ import { mix } from "mixwith";
 import * as THREE from "three";
 
 import { ATOM_GROUP_NAME } from "./enums";
+import { AtomLabelsMixin } from "./mixins/atomLabels";
 import { AtomsMixin } from "./mixins/atoms";
 import { BondsMixin } from "./mixins/bonds";
 import { BoundaryMixin } from "./mixins/boundary";
 import { CellMixin } from "./mixins/cell";
 import { ControlsMixin } from "./mixins/controls";
-// eslint-disable-next-line import/no-cycle
-import { LabelsMixin } from "./mixins/labels";
 import { MeasurementMixin } from "./mixins/measurement";
 import { RepetitionMixin } from "./mixins/repetition";
 import SETTINGS from "./settings";
-// eslint-disable-next-line import/no-cycle
-import { saveImageDataToFile } from "./utils";
+import { saveImageDataToFile } from "./utils/screenshot";
 
 const TV3 = THREE.Vector3;
 const TCo = THREE.Color;
@@ -243,7 +241,7 @@ export class Wave extends mix(WaveBase).with(
     RepetitionMixin,
     ControlsMixin,
     BoundaryMixin,
-    LabelsMixin,
+    AtomLabelsMixin,
     MeasurementMixin,
 ) {
     /**
